streamlit_app/requirements.txt
streamlit>=1.35.0
databricks-sql-connector>=3.6.0
databricks-sdk>=0.38.0
openai>=1.60.0
pandas>=2.0.0
beautifulsoup4>=4.11.0
requests>=2.31.0
